# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: rosshield API
  version: 0.0.1
  summary: ROS2 Fleet Security Audit Platform API
  description: |
    Skeleton spec for Phase 0. Contract-first — server handlers, client SDKs,
    and CLI command shapes are generated from this document. See
    `docs/design/05-api-and-auth.md` for the full design.

    Status: v1 under design. No operation here is considered stable.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:8080
    description: Local desktop / development
  - url: https://{host}
    description: On-prem or appliance
    variables:
      host:
        default: rosshield.example.internal
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      security: []
      summary: Local credential login
      description: |
        Dual mode (C6):
          - `X-Cookie-Auth: true` → refresh token은 응답 본문에서 제외되고
            HttpOnly `rosshield_refresh` cookie로 송출 (Web Console).
          - 헤더 부재 → 본문에 refresh token 포함 (CLI legacy 호환).
      x-status: todo
      parameters:
        - in: header
          name: X-Cookie-Auth
          schema: { type: string, enum: ["true"] }
          description: Set to `true` to receive refresh token via HttpOnly cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              schema: { type: string }
              description: HttpOnly refresh cookie (only when X-Cookie-Auth is true).
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string, description: "Omitted in cookie mode." }
                  user:
                    type: object
                    properties:
                      id:          { type: string }
                      email:       { type: string }
                      displayName: { type: string }
                      tenantId:    { type: string }
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshAuth
      security: []
      summary: Rotate access token using refresh token
      description: |
        Dual mode (C6): cookie 우선 → 본문 fallback.
          - Web Console (`X-Cookie-Auth: true`) → HttpOnly cookie에서 refresh 추출,
            새 refresh를 다시 cookie로 송출(rotation), 본문에는 access만.
          - CLI → 본문 `refreshToken` 사용, 본문에 새 refresh 포함.
        실패(만료/취소/재사용 탐지)는 모두 401 + cookie clear.
      x-status: todo
      parameters:
        - in: header
          name: X-Cookie-Auth
          schema: { type: string, enum: ["true"] }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string, description: "Legacy mode (CLI)." }
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              schema: { type: string }
              description: New HttpOnly refresh cookie (cookie mode only).
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string, description: "Omitted in cookie mode." }
        "401":
          description: Refresh failed (expired/revoked/reuse-detected)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logoutAuth
      security: []
      summary: Revoke refresh token and clear cookie
      description: |
        Idempotent. Clears `rosshield_refresh` cookie regardless of token presence.
        Cookie 또는 본문 `refreshToken`에서 토큰을 추출해 도메인 layer가 revoke.
      x-status: todo
      parameters:
        - in: header
          name: X-Cookie-Auth
          schema: { type: string, enum: ["true"] }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "204":
          description: No Content (cookie cleared)
          headers:
            Set-Cookie:
              schema: { type: string }
              description: Empty/expired refresh cookie.
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/auth/me:
    get:
      tags: [auth]
      operationId: getCurrentSession
      summary: Current session info
      x-status: todo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/tenants/current:
    get:
      tags: [tenant]
      operationId: getCurrentTenant
      summary: Current tenant
      x-status: todo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/robots:
    get:
      tags: [robot]
      operationId: listRobots
      summary: List robots
      x-status: todo
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Sort"
        - { in: query, name: fleetId, schema: { type: string } }
        - { in: query, name: tag,     schema: { type: string } }
        - { in: query, name: q,       schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRobotsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [robot]
      operationId: createRobot
      summary: Register a robot
      description: |
        Accepts plaintext credentials (memory-only) — domain layer wraps via KEK→DEK.
        Response excludes plaintext credentials (Robot meta + credentialId only).
        Conflicts (409): duplicate name within fleet, duplicate host:port.
      x-status: todo
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRobotRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRobotResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/robots/{robotId}:
    parameters:
      - { in: path, name: robotId, required: true, schema: { type: string, pattern: "^ro_[A-Z0-9]+$" } }
    get:
      tags: [robot]
      operationId: getRobot
      summary: Get a robot
      x-status: todo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/scans:
    get:
      tags: [scan]
      operationId: listScans
      summary: List scan sessions
      x-status: todo
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - { in: query, name: status,  schema: { type: string } }
        - { in: query, name: fleetId, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [scan]
      operationId: createScan
      summary: Start a new scan session
      description: |
        Phase 1 Stage B: inserts session in `pending` state — Orchestrator(scanrun)
        start is deferred to a later Stage. `total` is optional (orchestrator computes
        it; current Stage accepts external value).
      x-status: todo
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartScanRequest"
      responses:
        "201":
          description: Created (pending session)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSession"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/scans/{sessionId}/progress:
    parameters:
      - { in: path, name: sessionId, required: true, schema: { type: string } }
    get:
      tags: [scan]
      operationId: streamScanProgress
      summary: Stream scan progress (WebSocket upgrade)
      description: |
        WebSocket upgrade endpoint — OpenAPI 3.0/3.1 has no first-class WebSocket
        modeling, so this entry is documentation-only. The HTTP GET handler upgrades
        the connection (`coder/websocket.Accept`) and pushes JSON frames matching
        `ScanProgressMessage` schema (see `components/schemas/ScanProgressMessage`).
        See `internal/api/handlers/scan_ws.go` for the wire protocol.

        Subscribes to `scan.progress` and `scan.completed` events for the given
        sessionId within the caller tenant scope. Emits one `completed` frame
        (status=completed/failed/cancelled) and closes.

        Auth: Authorization Bearer is required; browser WebSocket API cannot set
        custom headers — Phase 1 assumes same-origin. Phase 2 will explore
        `?access_token=` query token or session cookie.
      x-status: phase1
      x-protocol: websocket
      responses:
        "101":
          description: Switching Protocols (WebSocket upgrade)
        "401":
          description: Missing/invalid bearer token
        "503":
          description: Event bus not configured
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/reports:
    get:
      tags: [reporting]
      operationId: listReports
      summary: List reports
      description: |
        Lists generated reports for the current tenant. PDF body is not included —
        only metadata + signature presence flag. Filter by `sessionId` to scope to a
        single scan.
      x-status: phase1
      parameters:
        - { in: query, name: sessionId, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListReportsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/reports/{reportId}:verify:
    post:
      tags: [reporting]
      operationId: verifyReport
      summary: Verify the signature of a report
      x-status: todo
      parameters:
        - { in: path, name: reportId, required: true, schema: { type: string } }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/license:
    get:
      tags: [tenant]
      operationId: getLicenseInfo
      summary: License metadata (E24)
      description: |
        Returns the active license edition (community|enterprise), expiry, active
        features, and quotas. Token and signature are NEVER exposed — only metadata
        the operator needs. License unset → community defaults.
      x-status: phase3
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  edition:    { type: string, enum: [community, enterprise] }
                  issuedTo:   { type: string }
                  issuedAt:   { type: string, format: date-time }
                  expiresAt:  { type: string, format: date-time }
                  expired:    { type: boolean }
                  features:   { type: array, items: { type: string } }
                  quotas:
                    type: object
                    properties:
                      robotsMax:       { type: integer, description: "0 = unlimited" }
                      scansPerDay:     { type: integer, description: "0 = unlimited" }
                      llmTokensPerDay: { type: integer, description: "0 = unlimited" }
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/audit/head:
    get:
      tags: [audit]
      operationId: getAuditHead
      summary: Current hash-chain head
      description: |
        Returns tenant-scoped chain head (Seq + Hash). When no audit entries exist,
        returns genesis head (Seq=0, Hash=zero) — Web UI renders "초기 상태".
      x-status: phase1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditHead"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/audit/verify:
    post:
      tags: [audit]
      operationId: verifyAuditChain
      summary: Verify a range of the audit chain (external verifier)
      x-status: todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from: { type: string, format: date-time }
                to:   { type: string, format: date-time }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        default:
          $ref: "#/components/responses/ErrorResponse"

  # E17 (Phase 2) — Insight 도메인 표면.
  /api/v1/insights:
    get:
      tags: [insight]
      operationId: listInsights
      summary: List active (non-dismissed) insights
      x-status: phase2
      parameters:
        - $ref: "#/components/parameters/Limit"
        - { in: query, name: kind,     schema: { type: string, enum: [drift, anomaly, peer, root_cause, prediction] } }
        - { in: query, name: severity, schema: { type: string, enum: [info, low, medium, high, critical] } }
        - { in: query, name: robotId,  schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListInsightsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/insights/{insightId}:dismiss:
    post:
      tags: [insight]
      operationId: dismissInsight
      summary: Mark an insight as dismissed
      x-status: phase2
      parameters:
        - { in: path, name: insightId, required: true, schema: { type: string, pattern: "^ins_[A-Z0-9]+$" } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insight"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/fleets/{fleetId}/insights:run:
    post:
      tags: [insight]
      operationId: runFleetInsights
      summary: Trigger insight detectors for a fleet (drift/anomaly/peer)
      description: |
        Manually runs the insight pipeline for the given fleet. E19 will subscribe
        this to scan.completed events for automation; this endpoint remains for
        on-demand recompute.
      x-status: phase2
      parameters:
        - { in: path, name: fleetId, required: true, schema: { type: string } }
      responses:
        "200":
          description: Detectors ran (may produce 0 insights if no history)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunInsightsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"

  # E17 (Phase 2) — Compliance 도메인 표면.
  /api/v1/compliance/profiles:
    get:
      tags: [compliance]
      operationId: listComplianceProfiles
      summary: List compliance profiles for the current tenant
      x-status: phase2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListComplianceProfilesResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [compliance]
      operationId: createComplianceProfile
      summary: Activate a compliance framework
      x-status: phase2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [framework, frameworkVersion]
              properties:
                framework:        { type: string, enum: [isms-p, iso27001-2022, nist-800-53-rev5] }
                frameworkVersion: { type: string }
                enabled:          { type: boolean, default: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceProfile"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/compliance/profiles/{profileId}/snapshots:
    parameters:
      - { in: path, name: profileId, required: true, schema: { type: string, pattern: "^cp_[A-Z0-9]+$" } }
    get:
      tags: [compliance]
      operationId: listComplianceSnapshots
      summary: List snapshots for a profile (created_at DESC)
      x-status: phase2
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListComplianceSnapshotsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [compliance]
      operationId: generateComplianceSnapshot
      summary: Generate a snapshot from a scan session
      x-status: phase2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId: { type: string }
      responses:
        "201":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceSnapshot"
        default:
          $ref: "#/components/responses/ErrorResponse"

  # E16/E19-3 (Phase 2) — Advisor 도메인 표면 (LLM 옵트인).
  #
  # LLM provider가 noop이면 모든 ask는 503 Service Unavailable (advisor.ErrAdvisorDisabled).
  # 활성화 방법: 서버 시작 시 `--llm-provider=ollama|=anthropic`.
  #
  # 핸들러는 chi 직접 mount (gen.ServerInterface 미구현, gen.Unimplemented가 자동 501).
  # spec은 문서·계약 목적으로만 정의.
  /api/v1/advisor/conversations:ask:
    post:
      tags: [advisor]
      operationId: askAdvisor
      summary: Ask the advisor a natural-language question (opt-in)
      description: |
        Starts a new conversation (when conversationId is empty) or appends to an
        existing one. The orchestrator runs an LLM/tool dispatch loop and returns
        the final answer plus the turns produced in this request.

        Returns 503 if the LLM provider is disabled (noop).
      x-status: phase2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                conversationId: { type: string, description: "Empty → new conversation" }
                question:       { type: string, minLength: 1 }
                maxToolCalls:   { type: integer, minimum: 0, maximum: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvisorAskResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/advisor/conversations:
    get:
      tags: [advisor]
      operationId: listAdvisorConversations
      summary: List the current user's conversations (updated_at DESC)
      x-status: phase2
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAdvisorConversationsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/advisor/conversations/{conversationId}:
    parameters:
      - { in: path, name: conversationId, required: true, schema: { type: string } }
    get:
      tags: [advisor]
      operationId: getAdvisorConversation
      summary: Get a conversation with all its turns (sequence ASC)
      x-status: phase2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvisorConversationDetail"
        default:
          $ref: "#/components/responses/ErrorResponse"

  # E23-C (Phase 3) — Webhook CRUD HTTP 표면.
  # Secret은 응답에서 마지막 4자만 노출 (`secretLast4`). 평문 secret은 PUT/POST 입력 전용.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by `/api/v1/auth/login` or OIDC/SAML callback
    apiKeyAuth:
      type: http
      scheme: bearer
      description: "API key in `Authorization: Bearer fg_live_...` form"
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
    Sort:
      in: query
      name: sort
      description: "Field name; prefix with `-` for descending."
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      description: ULID or UUID. Same key within 24h returns the original response.
      schema: { type: string, minLength: 16, maxLength: 64 }
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    Envelope:
      type: object
      required: [ok]
      properties:
        ok:    { type: boolean, const: true }
        value: {}
        meta:  { $ref: "#/components/schemas/Meta" }
...
    ListRobotsResponse:
      type: object
      required: [robots]
      properties:
        robots:
          type: array
          items: { $ref: "#/components/schemas/Robot" }
...
    CreateRobotResponse:
      type: object
      required: [robot, credentialId]
      properties:
        robot:        { $ref: "#/components/schemas/Robot" }
        credentialId: { type: string }

    # ─── Scan ─────────────────────────────────────────────────────────────
...
    ListReportsResponse:
      type: object
      required: [reports]
      properties:
        reports:
          type: array
          items: { $ref: "#/components/schemas/Report" }

    # ─── Insight ──────────────────────────────────────────────────────────
...